fix(apps): one row per app, and routing that survives new streams
Firefox showed up as several rows in the Apps tab and kept losing its
channel. Three separate causes:

- The apps list and the mixer's per-channel app list rendered one entry
  per stream, and a browser opens one per playing tab. Both now group by
  stream identity, so an app is one row carrying all of its streams, and
  routing or setting the volume on that row applies to every one of them.

- Auto-routing only ran inside the frontend's stream poll, which stops
  while the window is hidden in the tray (TD-009). A stream opened
  meanwhile landed on the default sink and stayed there. Enforcement now
  runs on a backend ticker that never sleeps; the frontend poll is only
  about what it draws.

- The "already handled" ledger was keyed on PipeWire node ids, which are
  recycled within seconds - an app closing one stream and opening another
  between two polls inherited the old entry and was never routed
  (observed here: node 247 across serials 94848, 95924 and 96615). It is
  keyed on object.serial now, which is never reused.

The ticker runs at 200ms on the native backend, so a browser rebuilding
its stream on a new video is back on its channel in ~180ms rather than
up to 2s. A check there costs ~150us (a channel round-trip to the
PipeWire thread, answered from the node map it already keeps), measured
at 0.15% of a core for the enforcement thread. The pactl fallback keeps
the 2s cadence and the poll-dedup: it forks two processes per check
(~13ms).

+++ src-tauri/src/mixer/state.rs
@@ -42,10 +42,13 @@ pub struct MixerState {
     pub buses: crate::persistence::buses::Buses,
     /// App preferences (device naming etc.), persisted to disk.
     pub prefs: crate::persistence::prefs::Prefs,
-    /// Stream indices already considered for auto-routing this session.
-    /// Each stream is enforced once, on first sight, so a user moving a
-    /// stream elsewhere (here or in pavucontrol) isn't fought every poll.
-    pub auto_routed: HashSet<u32>,
+    /// Streams already considered for auto-routing this session, by
+    /// `object.serial`. Each stream is enforced once, on first sight, so a
+    /// user moving a stream elsewhere (here or in pavucontrol) isn't fought
+    /// every poll. Serials, not node ids: PipeWire recycles node ids fast
+    /// enough that an app closing one stream and opening another between two
+    /// polls would inherit the old entry and never get routed.
+    pub auto_routed: HashSet<u64>,
 }
 
 impl MixerState {
